// frontend/src/shared/api/generated/models/document.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Databricks Apps FastAPI Starter API
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentStatus } from './documentStatus';

export interface Document {
  addedAt: string;
  id: string;
  name: string;
  projectId?: string | null;
  size: number;
  status: DocumentStatus;
  type: string;
}