// frontend/src/shared/api/generated/models/documentStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Databricks Apps FastAPI Starter API
 * OpenAPI spec version: 0.1.0
 */

export type DocumentStatus = typeof DocumentStatus[keyof typeof DocumentStatus];


export const DocumentStatus = {
  pending: 'pending',
  ingested: 'ingested',
  error: 'error',
} as const;